% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_coded}
\alias{data_coded}
\title{data_coded}
\format{
A data frame with 281 rows and 12 variables:
\describe{
\item{\code{name}}{character eHRAF culture name}
\item{\code{latitude}}{double Estimated population latitude}
\item{\code{longitude}}{double Estimated population longitude}
\item{\code{region}}{character eHRAF Region}
\item{\code{subregion}}{character eHRAF Subregion}
\item{\code{age.of.learner}}{character Catigorical age estimation of learner}
\item{\code{domain}}{character Social domain of learning}
\item{\code{mode}}{character Mode of social learning}
\item{\code{process}}{character Process of social learning}
\item{\code{sex}}{character Gender of learner(s)}
\item{\code{coding_ID}}{double Internally assigned ID for coding}
\item{\code{text_record_ID}}{integer Associated text ID}
}
}
\source{
\url{https://doi.org/10.1007/978-4-431-55997-9_2}
}
\usage{
data_coded
}
\description{
Researcher-coded social learning measures at the instance level
}
\keyword{datasets}
