% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_culture}
\alias{data_culture}
\title{data_culture}
\format{
A data frame with 27 rows and 10 variables:
\describe{
\item{\code{culture_id}}{character eHRAF-given OWC ID for culture.}
\item{\code{name}}{character eHRAF-given society name.}
\item{\code{region}}{character eHRAF-given continental region.}
\item{\code{subregion}}{character eHRAF-given continental sub-region.}
\item{\code{subsistence_type}}{character eHRAF-given subsistence type.}
\item{\code{n_pages}}{double Total number of pages in document.}
}
}
\source{
\url{https://ehrafworldcultures.yale.edu/ehrafe/}
}
\usage{
data_culture
}
\description{
Culture-level metadata
}
\keyword{datasets}
