% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_bias}
\alias{data_bias}
\title{DATASET_TITLE}
\format{
A data frame with 85 rows and 10 variables:
\describe{
\item{\code{document_id}}{character COLUMN_DESCRIPTION}
\item{\code{document_publication_date}}{integer COLUMN_DESCRIPTION}
\item{\code{document.type}}{character COLUMN_DESCRIPTION}
\item{\code{avg.fielddate}}{integer COLUMN_DESCRIPTION}
\item{\code{field.date}}{character COLUMN_DESCRIPTION}
\item{\code{coverage.date}}{character COLUMN_DESCRIPTION}
\item{\code{document_page_count}}{integer COLUMN_DESCRIPTION}
\item{\code{source.evaluation}}{character COLUMN_DESCRIPTION}
\item{\code{female_coauthorship}}{integer COLUMN_DESCRIPTION}
\item{\code{culture_id}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://ehrafworldcultures.yale.edu/ehrafe/}
}
\usage{
data_bias
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
