% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_culture}
\alias{data_culture}
\title{data_culture}
\format{
A data frame with 27 rows and 10 variables:
\describe{
\item{\code{culture_id}}{character The unique identifier for culture. OWC ID, given by HRAF.}
\item{\code{OWC Code}}{character The unique identifier for culture. OWC ID, given by HRAF.}
\item{\code{Name}}{character HRAF given society name.}
\item{\code{Region}}{character HRAF given continental region.}
\item{\code{Subregion}}{character HRAF given continental sub-region.}
\item{\code{Subsistence Type}}{character HRAF given subsistence type.}
\item{\code{SCCS}}{double COLUMN_DESCRIPTION}
\item{\code{SCCS_id}}{character COLUMN_DESCRIPTION}
\item{\code{N_docs}}{double COLUMN_DESCRIPTION}
\item{\code{N_pages}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
data_culture
}
\description{
Culture-level metadata.
}
\keyword{datasets}
